Shows prev/next links for mobile view
Why
---
Pagination on small screens only shows the last and first link in the
pagination. With the way it's currently set up, that means the first
page and last page only.

This change addresses the need by
---------------------------------
Remove last and first page links so that next and previous page links
show up on small screens

Side effects
------------
No longer have easy access to last and first page in pagination