Stops bug with syncing before tweet archived
Why
---
If the sheet sync is run before the tweet has been downloaded, it throws
an error expecting the tweet to exist.

This change addresses the need by
---------------------------------
Download all tweets before trying to sync to the sheet. Also resync if
the listen URL has been updated and is currently blank