Check for reply tweet when syncing to sheet
Why
---
The tweet that is usually saved on the spreadsheet is the reply to the
announcement tweet. This causes the order of the tweets to be off when
the tweet review is synced since announcment tweet should be parent of
the tweet review.

This change addresses the need by
---------------------------------
Check to see if tweet has reply and use that for tweet review if present