Adds user that is created from twitter user
Why
---
Real names are used on the JulySoundcheck spreadsheet which don't always
align to the twitter name. There needs to be a way to save the name that
is used on the spreadsheet.

This change addresses the need by
---------------------------------
Creates a user model that contains twitter details as well as real name.
Create profile edit page to update this real name as well as menu in top
bar to access it.