Fixes bug downloading non-tweet reviews
Why
---
Non-tweet reviews were being parsed with a reviewer set in the
attributes. This field was no longer being used since a user is found or
created directly from the sheet row with a different parser.

This change addresses the need by
---------------------------------
Removes the reviewer parsed attribute for non-tweet reviews and wrote
feature test to make sure these type of reviews are synced correctly